Move class `Persistable` into `gc3libs/persistence/serialization.py`.
Since class `Persistable` has no dependencies, and neither does any
other part of `gc3libs.persistence.serialization`, moving it allows
one to import class `Persistable` from other parts of GC3Libs (e.g.,
`gc3libs.url`) without causing circular dependency errors.
